Renamed roles to userRoles
